Moved logging function to Lib directory and refactored code.
